+++ lib/element/dropdown_btn_element.dart
@@ -14,6 +14,7 @@ import 'spwml_element.dart';
 class DropdownBtnElement extends SpWMLElement {
   final StructureElementChildren children;
   final ShowMenuBtnElementParams menuBtnParam;
+  final DropDownBtnElementParams ddBtnParam;
 
   DropdownBtnElement(
       int serial,
@@ -24,7 +25,8 @@ class DropdownBtnElement extends SpWMLElement {
       int lineEnd,
       SpWMLFontStyle style,
       this.children,
-      this.menuBtnParam)
+      this.menuBtnParam,
+      this.ddBtnParam)
       : super(serial, EnumSpWMLElementType.dropdownBtn, param, text,
             parentSerial, lineStart, lineEnd, style);
 
@@ -37,7 +39,8 @@ class DropdownBtnElement extends SpWMLElement {
       int lineEnd,
       SpWMLFontStyle style,
       this.children,
-      this.menuBtnParam)
+      this.menuBtnParam,
+      this.ddBtnParam)
       : super.convert(serial, EnumSpWMLElementType.dropdownBtn, param, text,
             parentSerial, lineStart, lineEnd, style);
 
@@ -83,13 +86,23 @@ class DropdownBtnElement extends SpWMLElement {
       return _DropDownElementWidget(
           menus,
           menuBtnParam,
+          ddBtnParam,
           icon,
           param.containsKey(EnumSpWMLElementParam.iconSize)
               ? param[EnumSpWMLElementParam.iconSize]
               : null);
     } else {
-      return _DropDownElementWidget(menus, menuBtnParam, null, null);
+      return _DropDownElementWidget(
+          menus, menuBtnParam, ddBtnParam, null, null);
     }
+  }
+
+  /// (en)Set initial index of dropdown items.
+  ///
+  /// (ja)ドロップダウンアイテムの初期インデックスを設定します。
+  /// * [index] : The index. The first child element is counted as 0.
+  void setInitialIndex(int index) {
+    ddBtnParam.index = index;
   }
 
   /// (en)Set menus callback.
@@ -103,20 +116,19 @@ class DropdownBtnElement extends SpWMLElement {
 
 class _DropDownElementWidget extends StatefulWidget {
   final List<DropdownMenuItem<int>> menus;
-  final ShowMenuBtnElementParams ddBtnParam;
+  final ShowMenuBtnElementParams menuBtnParam;
+  final DropDownBtnElementParams ddBtnParam;
   final Icon? icon;
   final double? iconSize;
 
   const _DropDownElementWidget(
-      this.menus, this.ddBtnParam, this.icon, this.iconSize);
+      this.menus, this.menuBtnParam, this.ddBtnParam, this.icon, this.iconSize);
 
   @override
   _DropDownElementWidgetState createState() => _DropDownElementWidgetState();
 }
 
 class _DropDownElementWidgetState extends State<_DropDownElementWidget> {
-  int _selectedItem = 0;
-
   @override
   void initState() {
     super.initState();
@@ -128,11 +140,11 @@ class _DropDownElementWidgetState extends State<_DropDownElementWidget> {
       icon: widget.icon,
       iconSize: widget.iconSize ?? 24.0,
       items: widget.menus,
-      value: _selectedItem,
+      value: widget.ddBtnParam.index,
       onChanged: (int? v) => {
         setState(() {
           if (v != null) {
-            _selectedItem = v;
+            widget.ddBtnParam.index = v;
           }
         }),
       },